pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "deepagents>=0.5.3",
    "langchain>=1.2.15",
    "langchain-core>=1.3.0",
    "langchain-deepseek>=1.0.1",
    "langgraph>=1.1.9",
    "python-dotenv>=1.2.2",
]
